+++ lib/src/parsers/url_parser.dart
@@ -12,17 +12,7 @@ dynamic urlParser(Parser parser, dynamic dom) {
   if (selector != null) {
     String? attribute = selector.attributes['href'];
     if (attribute != null) {
-      String prepend = "";
-      if (parser.optional != null && parser.optional!.prepend != null) {
-        prepend = parser.optional!.prepend!;
-      }
-      if (attribute.startsWith('/')) {
-        return prepend + attribute.toString();
-      }
-      if (attribute.contains('http')) {
-        return prepend + attribute.toString();
-      }
-      return null;
+      return attribute.toString();
     }
   }
   return null;